Trekk ut Hamburgermeny til egen komponent
